1.10.10 Task: Added Tests for CompareDatabase & PatchDatabase

+++ src/main/de/jClipCorn/database/elementProps/impl/EOptShortProp.java
@@ -13,7 +13,7 @@ public class EOptShortProp extends EOptProperty<Short> {
 
 	@Override
 	protected String serializeInnerToString(Short v) {
-		return String.valueOf(v);
+		return String.valueOf((short)v);
 	}
 
 	@Override
